Introduce calculator FIDL definitions
These types allow a client and a server to communicate with each other
over FIDL to perform unary and binary mathematical operations. The main
entry point for these operations is the Calculator protocol.

Such types could be used to create an interactive calculator. The client
of the Calculator protocol could be a graphical Fuchsia component that
offloads its arithmetic computations to a server component.

Change-Id: Ie92475bebfdf35b26d8ad0261594901455a6ba75