Add basic arithmetic operations
This is the initial commit of the calculator engine. These operations
are implemented as ordinary C++ functions that work synchronously and
return a value.

By implementing these as a set of sources with no reference to Fuchsia
concepts, it's possible to write unit tests against these functions
where the tests run on the host machine. This would not be possible if
the implementation referenced FIDL in any way.

Change-Id: I3c8c9faaa50a78e00a4ee26ab598d3f232027a8e