Add device tests for completed Engine package
Unlike the tests added in I3c8c9faa, the tests added in this change
run on an actual Fuchsia device (or emulator). They verify via the
Calculator protocol that the Calculator Engine component behaves
correctly.

Note that the main difference between these tests and the ones in change
I3c8c9faaa are the following:

* Where they run: these tests run on device
* What interface they use: these tests call into the FIDL service

Change-Id: I3fa3b0766454304068af321ca18e2670096c294d