Fix gsutil detection by normalizing sdk/.. before the directory exists
When starting for the first time, the sdk directory does not exist. So if
you try and search for sdk/../bin/gsutil, it will always fail. By using
os.path.normpath, we cancel out sdk/.. and since we are going up a directory
we don't need the sdk directory to actually exist yet.

Fixed: 42594
Change-Id: I94347a5f6d9d2838a08f45931c651f862e4b0362