Introduce top-level clang-format style options
The presence of this file standardizes the behaviour of clang-format.
These formatting rules are the same ones used in-tree in Fuchsia.

This file also allows the C/C++ extension in Visual Studio Code to
operate correctly. That extension always passes an explicit value for
the style flag, so it's not possible to use the bundled copy of
clang-format with its default settings. Instead, the extension can be
left with a "style" value of "file", and this will cause clang-format to
look for the .clang-format file introduced in this change.

Change-Id: I74e76f1e6fcdfa77e695517978277bfdb9b5ce43