Make md5sum do hash validation
Previously, a newly downloaded Core SDK tarball would be hashed, and
this hash result would be stored in $CHECKSUM_FILE. Future downloads
would be started if this $CHECKSUM_FILE didn't match a one that was
constructed at runtime. Unfortunately, the constructed $CHECKSUM_FILE
doesn't always exactly match the whitespace in the output of md5sum, and
this would case unnecessary redownloads.

This change instead asks the md5sum command to perform the hash
validation. This allows whitespace to vary as long as the individual
fields in the checksum file match their expected values.

Change-Id: Icc338bfd12d6d449791b987f2fe7c346c16e5a84