[gn sdk]Updating to the new GN SDK structure.
The GN SDK tarball is unzipped into //third_party/fuchsia-sdk.

The effect of this is the flattening of the /sdk/ directory. This
was interfering with other projects that were downloading the tarball
into a sub directory, causing an ugly /sdk/sdk/ some of the paths.

Another addition is the propagation of the metadata from the IDK.
All files in the SDK are accounted for in the meta/manifest.json and
related files.

Change-Id: Ic1af7a9ff816d2ead504c0ed916c81fa15e93f17