Add update SDK script
This CL adds a update SDK script to allow maintainers to easily
update the project to the latest version. Running
`scripts/update-fuchsia-sdk.sh` will download the SDK from GCS
extract it to a temporary directory, remove the existing SDK and
replace it with the latest available SDK.

The purpose of adding this script is to make it easy to update the
SDK with the latest changes until a roller can be constructed to
update the SDK to the latest version in a automated way.

Change-Id: Ice8c33c78441ba87423b14044da4ec7d2fa602dc