[samples] Specify exact versions of CIPD downloads to get hermetic builds
Previously we just downloaded "latest" which could change or be broken
at any time. By specifying an exact version and putting this into git,
we are now able to replicate the same build tools at any time in the
future. The versions used were taken from $FUCHSIA_ROOT/integration/prebuilts
which are known to work in-tree, but should not need to be adjusted in
the future since these tools are quite stable.

Change-Id: Ib0edd0a5f8b55106dc74e7cdbe4db31f7e405868