[gn sdk samples] Adding CIPD packaging rules
This adds the symbol_archive template and demonstrates how to use it
to create a symbol archive for the hello_world project.

There are also changes to the toolchain to support building stripped
binaries.  The toolchain is external to the fuchsia SDK, the other
changes are part of the fuchsia sdk, and will be merged in the fuchsia
source tree, but they are included here for instructional use and
prototyping.

This also adds the buildidtool which is used to create the .build-id
symbol directories.

Change-Id: If74ba455ba4c4f7efa173f449c128115acc8b64b
Reviewed-on: https://fuchsia-review.googlesource.com/c/samples/+/398043
Reviewed-by: Wayne Piekarski <waynepie@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>